refactor: break taskLearning metrics<->routing static import cycle (#5916)
Move resetTaskTypeLearning, removeTaskTypeFromLearningData, and
purgeEnvironmentalFailuresForType into a new store-only leaf
(taskLearning/reset.js). routing.js imports the leaf instead of metrics.js,
so the two-module ESM cycle is gone; metrics.js and the barrel re-export the
moved symbols for back-compat. Drops the #5916 entry from
KNOWN_CYCLIC_COMPONENTS.
