+++ server/services/serviceImportCycles.test.js
@@ -83,8 +83,6 @@ const KNOWN_CYCLIC_COMPONENTS = [
   { issue: 5918, members: ['pipeline/manuscriptFix.js', 'pipeline/manuscriptReview.js'] },
   // #5919 — the receive path reaches subscription state back through the barrel.
   { issue: 5919, members: ['sharing/peerSync.js', 'sharing/peerSyncReceive.js'] },
-  // #5916 — routing imports the metrics reset; metrics imports routing's predicates.
-  { issue: 5916, members: ['taskLearning/metrics.js', 'taskLearning/routing.js'] },
 ];
 
 // A component as one comparable string. Members arrive sorted from
